// Reclutamiento/ClientApp/src/app/shared/enums/type-event.enum.ts
export enum TypeEvent {
  RemovingRequisition = 'removingRequisition',
  SavingRequisition = 'savingRequisition',
  SavingRequisitionSendingToValidationRequisition = 'savingRequisitionSendingToValidationRequisition',
  SendingToValidationRequisition = 'sendingToValidationRequisition',
  ValidatingAuthorizationRequisition = 'validatingAuthorizationRequisition',
  AssignedRequisition = 'assignedRequisition',
  AddingCandidate = 'addingCandidate',
  AddingCandidateToTheRequisition = 'addingCandidateToTheRequisition',
  AssigningTernaToCandidateRequisition = 'assigningTernaToCandidateRequisition',
  LookingForACollaboratorByUsername = 'lookingForACollaboratorByUsername',
  SavingInterviewCandidateRequisition = 'savingInterviewTheCandidateRequisition',
  AddingGuestCandidateInterviewRequisition = 'addingGuestCandidateInterviewRequisition',
  SelectingSuitableCandidateRequisition = 'selectingSuitableCandidateRequisition',
  SavingCommentaryInterviewCandidateRequisition = 'savingCommentaryInterviewCandidateRequisition',
  SavingCompetencyKnowledgeInterviewCandidateRequisition = 'savingCompetencyKnowledgeInterviewCandidateRequisition',
  SavingCompetencyLeadershipInterviewCandidateRequisition = 'savingCompetencyLeadershipInterviewCandidateRequisition',
  SavingCompetencyInterviewCandidateRequisition = 'savingCompetencyInterviewCandidateRequisition',
  SavingFormatInterviewCandidateRequisition = 'savingFormatInterviewCandidateRequisition',
  SavingDocumentRequisition = 'savingDocumentRequisition',
  DeletingDocumentRequisition = 'deletingDocumentRequisition',
  SavingDocumentEconomicOfferRequisition = 'savingDocumentEconomicOfferRequisition',
  DeletingDocumentEconomicOfferRequisition = 'deletingDocumentEconomicOfferRequisition',
  SendingOfferRequisitionCandidate = 'sendingOfferRequisitionCandidate',
  AcceptingOfferRequisitionCandidate = 'acceptingOfferRequisitionCandidate',
  SendingDocumentationOfCandidateFile = 'sendingDocumentationOfCandidateFile',
  RespondingToTheDocumentationOfCandidateFile = 'respondingToTheDocumentationOfCandidateFile',
  SavingDateEntrySuitableCandidateRequisition = 'savingDateEntrySuitableCandidateRequisition',
  SavingNotificationAltaSuitableCandidateRequisition = 'savingNotificationAltaSuitableCandidateRequisition',
  SavingConfirmationAltaSuitableCandidateRequisition = 'savingConfirmationAltaSuitableCandidateRequisition',
  SendingSalaryOfferSuitableCandidateForTheRequisition = 'sendingSalaryOfferSuitableCandidateForTheRequisition',
  GettingPlaces = 'gettingPlaces',
  GettingAlias = 'gettingAlias',
  GettingJobInformationAdam = 'gettingJobInformationAdam',
  SavingTypeSearchForTheRequisition = 'savingTypeSearchForTheRequisition',
  AssociatingDomainUserWithNewCollaborator = 'associatingDomainUserWithNewCollaborator',
  VerifyingCandidateMail = 'verifyingCandidateMail',
  SavingCandidateComplement = 'savingCandidateComplement',
  SavingComplementCV = 'savingComplementCV',
  SavingDocumentCVCandidate = 'savingDocumentCVCandidate',
}